requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
requests==2.31.0
python-dateutil==2.8.2
bcrypt==4.0.1
PyJWT==2.8.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3
